Refactoring about category database
Add reference feature. Display flash messages in admin page.

+++ routes/admin.js
@@ -10,24 +10,23 @@ router.get('/', isAuth, (req, res, next) => {
 });
 
 router.get('/category', isAuth, async (req, res, next) => {
-  const categories = await Category.find();
+  const categories = await Category.find().populate('children');
   res.render('admin/category/index', { title: 'Category Manage - Admin page', categories: categories });
 });
 
 router.get('/category/new', isAuth, async (req, res, next) => {
-  const categories = await Category.find();
+  const categories = await Category.find().populate('children');
   res.render('admin/category/new', { title: 'Category Manage - Admin page', categories: categories });
 });
 
 router.get('/category/:id', isAuth, async (req, res, next) => {
-  const categories = await Category.find();
-  const category = await Category.findById(req.params.id);
+  const categories = await Category.find().populate('children');
+  const category = await Category.findById(req.params.id).populate('parent');
 
   res.render('admin/category/edit', { title: 'Category Manage - Admin page', categories: categories, category: category });
 });
 
 router.put('/category/:id', isAuth, async (req, res, next) => {
-  // TODO: 자식이 있는 카테고리 변경시 문제 있음. refence로 교체 요망.
   const category = await Category.findById(req.params.id);
 
   if (!category) {
@@ -36,7 +35,8 @@ router.put('/category/:id', isAuth, async (req, res, next) => {
   }
 
   category.name = req.body.category_title;
-  category.parent = req.body.category_parent;
+  const parent_id = await Category.findOne({ name: req.body.category_parent });
+  category.parent = parent_id;
 
   await category.save();
   req.flash('success', 'Successfully updated!');
@@ -44,27 +44,42 @@ router.put('/category/:id', isAuth, async (req, res, next) => {
 });
 
 router.delete('/category/:id', isAuth, async (req, res, next) => {
-  await Category.findOneAndRemove({ _id: req.params.id });
-  req.flash('success', 'Successfully deleted.');
-  res.redirect('/admin/category');
+  const category = await Category.findById(req.params.id).populate('children');
+  if (category.children.length !== 0) {
+    req.flash('error', 'There are children in that category. Please delete them first.');
+    res.redirect('/admin/category');
+  } else {
+    await Category.findOneAndRemove({ _id: req.params.id });
+    req.flash('success', 'Successfully deleted.');
+    res.redirect('/admin/category');
+  }
 });
 
 router.post('/category', isAuth, async (req, res, next) => {
   let category = await Category.findOne({ name: req.body.category_title });
-
   if (category) {
     req.flash('error', 'This category already exist.');
     return res.redirect('back');
-  } else {
-    const parent_id = await Category.findOne({ name: req.body.category_parent });
+  }
+  if (req.body.category_parent) {
+    let parent = await Category.findOne({ name: req.body.category_parent });
+    if (!parent) {
+      req.flash('error', 'Parent not exist.');
+      return res.redirect('back');
+    }
     category = new Category({
       name: req.body.category_title,
-      parent: parent_id
+      parent: parent._id
     });
-    await category.save();
-    req.flash('success', 'New category saved successfully.');
-    res.redirect('/admin/category');
+  } else {
+    category = new Category({
+      name: req.body.category_title,
+    });
   }
+
+  await category.save();
+  req.flash('success', 'New category saved successfully.');
+  res.redirect('/admin/category');
 });
 
 router.get('/user', isAuth, async (req, res, next) => {
